Backend/requirements.txt
fastapi[standard]
uvicorn 
python-dotenv

pymongo